:sparkles: split bits lib
