+++ src/nasal_builtin.h
@@ -54,11 +54,6 @@ var builtin_num(var*, gc&);
 var builtin_pop(var*, gc&);
 var builtin_str(var*, gc&);
 var builtin_size(var*, gc&);
-var builtin_u32xor(var*, gc&);
-var builtin_u32and(var*, gc&);
-var builtin_u32or(var*, gc&);
-var builtin_u32nand(var*, gc&);
-var builtin_u32not(var*, gc&);
 var builtin_time(var*, gc&);
 var builtin_contains(var*, gc&);
 var builtin_delete(var*, gc&);
@@ -84,10 +79,6 @@ var builtin_tell(var*, gc&);
 var builtin_readln(var*, gc&);
 var builtin_stat(var*, gc&);
 var builtin_eof(var*, gc&);
-var builtin_fld(var*, gc&);
-var builtin_sfld(var*, gc&);
-var builtin_setfld(var*, gc&);
-var builtin_buf(var*, gc&);
 var builtin_sleep(var*, gc&);
 var builtin_pipe(var*, gc&);
 var builtin_fork(var*, gc&);
