+++ src/nasal_codegen.cpp
@@ -15,6 +15,7 @@ void codegen::load_native_function_table(nasal_builtin_table* table) {
 void codegen::init_native_function() {
     load_native_function_table(builtin);
     load_native_function_table(math_lib_native);
+    load_native_function_table(bits_native);
     load_native_function_table(coroutine_native);
     load_native_function_table(flight_gear_native);
 }
